Include crow fly distance calculation in Helper class
